+++ pysbd/lang/common/standard.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 from pysbd.utils import Rule
-from pysbd.abbreviation_replacer import AbbreviationReplacer
-from pysbd.lang.common.ellipsis import EllipsisRules
 
 class Standard:
 
@@ -21,12 +19,6 @@ class Standard:
     ExtraWhiteSpaceRule = Rule(r'\s{3,}', ' ')
 
     SubSingleQuoteRule = Rule(r'&⎋&', "'")
-
-    EllipsisRules = EllipsisRules
-
-    class AbbreviationReplacer(AbbreviationReplacer):
-        SENTENCE_STARTERS = "A Being Did For He How However I In It Millions "
-        "More She That The There They We What When Where Who Why".split(" ")
 
     class Abbreviation(object):
         """Defines the abbreviations for each language (if available)"""
@@ -81,6 +73,27 @@ class Standard:
                FullWidthQuestionMark, MixedDoubleQE, MixedDoubleQQ, MixedDoubleEQ,
                MixedDoubleEE, LeftParens, RightParens, TemporaryEndingPunctutation,
                Newline]
+
+    class EllipsisRules(object):
+
+        # below rules aren't similar to original rules of pragmatic segmenter
+        # modification: spaces replaced with same number of symbols
+        # Rubular: http://rubular.com/r/i60hCK81fz
+        ThreeConsecutiveRule = Rule(r'\.\.\.(?=\s+[A-Z])', '☏☏.')
+
+        # Rubular: http://rubular.com/r/Hdqpd90owl
+        FourConsecutiveRule = Rule(r'(?<=\S)\.{3}(?=\.\s[A-Z])', 'ƪƪƪ')
+
+        # Rubular: http://rubular.com/r/YBG1dIHTRu
+        ThreeSpaceRule = Rule(r'(\s\.){3}\s', '♟♟♟♟♟♟♟')
+
+        # Rubular: http://rubular.com/r/2VvZ8wRbd8
+        FourSpaceRule = Rule(r'(?<=[a-z])(\.\s){3}\.($|\\n)', '♝♝♝♝♝♝♝')
+
+        OtherThreePeriodRule = Rule(r'\.\.\.', 'ƪƪƪ')
+
+        All = [ThreeSpaceRule, FourSpaceRule, FourConsecutiveRule,
+               ThreeConsecutiveRule, OtherThreePeriodRule]
 
     class ReinsertEllipsisRules(object):
         # below rules aren't similar to original rules of pragmatic segmenter
